backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyMuPDF
python-multipart
aiofiles
requests
sentence-transformers
numpy
faiss-cpu
nltk
redis
openai
python-jose
google-auth
requests
